+++ src/camera.rs
@@ -1,6 +1,21 @@
 use bevy::prelude::*;
 
+use super::state::AppState;
+
 use super::Ship;
+
+pub struct CameraPlugin;
+impl Plugin for CameraPlugin {
+    fn build(&self, app: &mut App) {
+        app.add_systems(Startup, setup);
+        app.add_systems(Update, follow_player.run_if(in_state(AppState::Active)));
+    }
+}
+
+fn setup(mut commands: Commands) {
+    // Spawns game camera
+    commands.spawn(Camera2dBundle::default());
+}
 
 pub fn follow_player(
     mut camera: Query<&mut Transform, (With<Camera>, Without<Ship>)>,